requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests
python-dotenv
PyYAML==6.0
